fix: change SurfaceArea type from decimal to int in CreatePropertyCommand and database schema for consistency

+++ src/Application/Commands/Property/CreatePropertyCommand.cs
@@ -13,7 +13,7 @@ public class CreatePropertyCommand : IRequest<CreatePropertyResponse>
     public required string Location { get; set; }
     public decimal Price { get; set; }
     public int Bedrooms { get; set; }
-    public decimal SurfaceArea { get; set; }
+    public int SurfaceArea { get; set; }
     public int? AgencyId { get; set; }
 }
